int-> int64_t

+++ tasks/word2vec/word2vec.cpp
@@ -2,10 +2,11 @@
 
 #include <vector>
 
-int ScalarProduct(const std::vector<int>& v1, const std::vector<int>& v2) {
-    int scalar_product = 0;
+int64_t ScalarProduct(const std::vector<int>& v1, const std::vector<int>& v2) {
+    int64_t scalar_product = 0;
     for (std::size_t i = 0; i < v1.size(); ++i) {
-        scalar_product += v1[i] * v2[i];
+        scalar_product += static_cast<int64_t>(v1[i]) *
+                          static_cast<int64_t>(v2[i]);
     }
     return scalar_product;
 }
@@ -14,9 +15,9 @@ std::vector<std::string> FindClosestWords(const std::vector<std::string>& words,
                                           const std::vector<std::vector<int>>& vectors) {
 
     std::vector<std::string> closest_words;
-    int max_scalar_product = INT32_MIN;
+    int64_t max_scalar_product = INT64_MIN;
     for (std::size_t i = 1; i < words.size(); ++i) {
-        int scalar_product = ScalarProduct(vectors[0], vectors[i]);
+        int64_t scalar_product = ScalarProduct(vectors[0], vectors[i]);
         if (scalar_product < max_scalar_product) {
             continue;
         } else if (scalar_product > max_scalar_product) {
